moved Download logic to the Restore commandblocks
