+++ Arius/Arius.Core/Commands/Restore/RestoreCommandBlocks.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Concurrent;
+using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Threading;
@@ -139,17 +140,22 @@ internal class DownloadBinaryBlock : ChannelTaskBlockBase<PointerFile>
         int maxDegreeOfParallelism)
         : base(loggerFactory: loggerFactory, sourceFunc: sourceFunc, maxDegreeOfParallelism: maxDegreeOfParallelism, onCompleted: onCompleted)
     {
-        this.fileService = fileService;
-        this.repo = repo;
-        this.options = options;
+        this.fileService      = fileService;
+        this.repo             = repo;
+        this.options          = options;
         this.chunkRehydrating = chunkRehydrating;
     }
 
-    private readonly ConcurrentDictionary<BinaryHash, TaskCompletionSource<BinaryFile>> restoredBinaries = new();
-    private readonly FileService fileService;
-    private readonly Repository repo;
+    private readonly FileService            fileService;
+    private readonly Repository             repo;
     private readonly IRestoreCommandOptions options;
-    private readonly Action chunkRehydrating;
+    private readonly Action                 chunkRehydrating;
+
+    private readonly ConcurrentDictionary<BinaryHash, TaskCompletionSource<BinaryFile>> restoredBinaries = new();
+
+    // For unit testing purposes
+    internal static bool RestoredFromLocal      { get; set; } = false;
+    internal static bool RestoredFromOnlineTier { get; set; } = false;
 
     protected override async Task ForEachBodyImplAsync(PointerFile pf, CancellationToken ct)
     {
@@ -214,7 +220,7 @@ internal class DownloadBinaryBlock : ChannelTaskBlockBase<PointerFile>
                 bool restored;
                 try
                 {
-                    restored = await repo.Binaries.TryDownloadAsync(pf.BinaryHash, FileSystemService.GetBinaryFileInfo(pf), options);
+                    restored = await TryDownloadAsync(pf.BinaryHash, FileSystemService.GetBinaryFileInfo(pf), options);
                 }
                 catch (Exception e)
                 {
@@ -281,7 +287,96 @@ internal class DownloadBinaryBlock : ChannelTaskBlockBase<PointerFile>
             pf.Delete();
     }
 
-    // For unit testing purposes
-    internal static bool RestoredFromLocal { get; set; } = false;
-    internal static bool RestoredFromOnlineTier { get; set; } = false;
+    /// <summary>
+    /// Download the given Binary with the specified options.
+    /// Start hydration for the chunks if required.
+    /// Returns null if the Binary is not yet hydrated
+    /// </summary>
+    private async Task<bool> TryDownloadAsync(BinaryHash bh, BinaryFileInfo target, IRestoreCommandOptions options, bool rehydrateIfNeeded = true)
+    {
+        var chs = await repo.Binaries.GetChunkListAsync(bh);
+        var chunks = chs.Select(ch => (ChunkHash: ch, ChunkBlob: repo.Chunks.GetChunkBlobByHash(ch, requireHydrated: true))).ToArray();
+
+        var chunksToHydrate = chunks
+            .Where(c => c.ChunkBlob is null)
+            .Select(c => repo.Chunks.GetChunkBlobByHash(c.ChunkHash, requireHydrated: false));
+        if (chunksToHydrate.Any())
+        {
+            chunksToHydrate = chunksToHydrate.ToArray();
+            //At least one chunk is not hydrated so the Binary cannot be downloaded
+            logger.LogInformation($"{chunksToHydrate.Count()} chunk(s) for '{bh.ToShortString()}' not hydrated. Cannot yet restore.");
+
+            if (rehydrateIfNeeded)
+                foreach (var c in chunksToHydrate)
+                    //hydrate this chunk
+                    await repo.Chunks.HydrateAsync(c);
+
+            return false;
+        }
+        else
+        {
+            //All chunks are hydrated  so we can restore the Binary
+            logger.LogInformation($"Downloading Binary '{bh.ToShortString()}' from {chunks.Length} chunk(s)...");
+
+            var p = await repo.Binaries.GetPropertiesAsync(bh);
+            var stats = await new Stopwatch().GetSpeedAsync(p.ArchivedLength, async () =>
+            {
+                await using var ts = target.OpenWrite(); // TODO add async 
+
+                // Faster version but more code
+                //if (chunks.Length == 1)
+                //{
+                //    await using var cs = await chunks[0].ChunkBlob.OpenReadAsync();
+                //    await CryptoService.DecryptAndDecompressAsync(cs, ts, options.Passphrase);
+                //}
+                //else
+                //{
+                //    var x = new ConcurrentDictionary<ChunkHash, byte[]>();
+
+                //    var t0 = Task.Run(async () =>
+                //    {
+                //        await Parallel.ForEachAsync(chunks,
+                //            new ParallelOptions() { MaxDegreeOfParallelism = 20 },
+                //            async (c, ct) =>
+                //            {
+                //                await using var ms = new MemoryStream();
+                //                await using var cs = await c.ChunkBlob.OpenReadAsync();
+                //                await CryptoService.DecryptAndDecompressAsync(cs, ms, options.Passphrase);
+                //                if (!x.TryAdd(c.ChunkHash, ms.ToArray()))
+                //                    throw new InvalidOperationException();
+                //            });
+                //    });
+
+                //    var t1 = Task.Run(async () =>
+                //    {
+                //        foreach (var (ch, _) in chunks)
+                //        {
+                //            while (!x.ContainsKey(ch))
+                //                await Task.Yield();
+
+                //            if (!x.TryRemove(ch, out var buff))
+                //                throw new InvalidOperationException();
+
+                //            await ts.WriteAsync(buff);
+                //            //await x[ch].CopyToAsync(ts);
+                //        }
+                //    });
+
+                //    Task.WaitAll(t0, t1);
+                //}
+
+                foreach (var (_, cb) in chunks)
+                {
+                    await using var cs = await cb.OpenReadAsync();
+                    await CryptoService.DecryptAndDecompressAsync(cs, ts, options.Passphrase);
+                }
+            });
+
+            logger.LogInformation($"Downloading Binary '{bh.ToShortString()}' of {p.ArchivedLength.GetBytesReadable()} from {chunks.Length} chunk(s)... Completed in {stats.seconds}s ({stats.MBps} MBps / {stats.Mbps} Mbps)");
+
+            return true;
+        }
+    }
+
+
 }
